article 순번 타입 long으로 변경

+++ src/main/java/kr/codesqaud/cafe/repository/JdbcArticleRepository.java
@@ -42,7 +42,7 @@ public class JdbcArticleRepository implements ArticleRepository {
         return new RowMapper<Article>() {
             @Override
             public Article mapRow(ResultSet rs, int rowNum) throws SQLException {
-                return new Article(rs.getString("writer"), rs.getString("title"), rs.getString("content"), rs.getInt("index"));
+                return new Article(rs.getString("writer"), rs.getString("title"), rs.getString("content"), rs.getLong("index"));
             }
         };
     }
